refactor: changed folder (module): Ingestor -> usxingestor
