+++ services/usxingestor/strongsingestor.py
