+++ services/usxingestor/translationnote.py
@@ -3,10 +3,10 @@ import re
 
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    from ingestor.chapter import Chapter
+    from usxingestor.chapter import Chapter
     from manager.logmanager import LogManager
 
-from ingestor.verse import Verse
+from usxingestor.verse import Verse
 
 #region Cases To Handle (both from source and for destination for both cross references and footnotes)
 # 2KI 6:31-7:20
